+++ adapters/BidMachine/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## BidMachine iOS Mediation Adapter Changelog
 
+#### Next Version
+- Upgrade BidMachine adapter and unit tests to Swift 6.
+
 #### Version 3.6.1.0 (In progress)
 
 #### [Version 3.6.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/bidmachine/BidMachineAdapter-3.6.0.0.zip)